Include singular APIs under same package
