+++ src/routes.py
@@ -6,29 +6,29 @@ from src.resources.resource_network import NetworkResource, NetworkResourceByNam
 from src.resources.resource_point import PointResource, PointResourceByName, PointResourceList
 from src.resources.resource_wires_plat import WiresPlatResource
 from src.source_drivers.generic.resources.device.device_plural import GenericDevicePlural
-from src.source_drivers.generic.resources.device.device_singular import GenericDeviceSingular
-from src.source_drivers.generic.resources.device.device_singular_by_name import GenericDeviceSingularByName
+from src.source_drivers.generic.resources.device.device_singular import GenericDeviceSingularByUUID, \
+    GenericDeviceSingularByName
 from src.source_drivers.generic.resources.network.network_plural import GenericNetworkPlural
-from src.source_drivers.generic.resources.network.network_singular import GenericNetworkSingular
-from src.source_drivers.generic.resources.network.network_singular_by_name import GenericNetworkSingularByName
+from src.source_drivers.generic.resources.network.network_singular import GenericNetworkSingularByUUID, \
+    GenericNetworkSingularByName
 from src.source_drivers.generic.resources.point.point_plural import GenericPointPlural
-from src.source_drivers.generic.resources.point.point_singular import GenericPointSingular
-from src.source_drivers.generic.resources.point.point_singular_by_name import GenericPointSingularByName
+from src.source_drivers.generic.resources.point.point_singular import GenericPointSingularByUUID, \
+    GenericPointSingularByName
 from src.source_drivers.generic.resources.point.point_value_writer import GenericUUIDPointValueWriter, \
     GenericNamePointValueWriter
 from src.source_drivers.modbus.resources.device.device_plural import ModbusDevicePlural
-from src.source_drivers.modbus.resources.device.device_singular import ModbusDeviceSingular
-from src.source_drivers.modbus.resources.device.device_singular_by_name import ModbusDeviceSingularByName
+from src.source_drivers.modbus.resources.device.device_singular import ModbusDeviceSingularByUUID, \
+    ModbusDeviceSingularByName
 from src.source_drivers.modbus.resources.mapping.mapping import MPGBPMappingResourceList, \
     MPGBPMappingResourceByGenericPointUUID, MPGBPMappingResourceByBACnetPointUUID, \
     MPGBPMappingResourceByModbusPointUUID, MPGBPMappingResourceByUUID
 from src.source_drivers.modbus.resources.network.network_plural import ModbusNetworkPlural
-from src.source_drivers.modbus.resources.network.network_singular import ModbusNetworkSingular
-from src.source_drivers.modbus.resources.network.network_singular_by_name import ModbusNetworkSingularByName
+from src.source_drivers.modbus.resources.network.network_singular import ModbusNetworkSingularByName, \
+    ModbusNetworkSingularByUUID
 from src.source_drivers.modbus.resources.point.point_plural import ModbusPointPlural
 from src.source_drivers.modbus.resources.point.point_poll import ModbusPointPollNonExisting, ModbusPointPoll
-from src.source_drivers.modbus.resources.point.point_singular import ModbusPointSingular
-from src.source_drivers.modbus.resources.point.point_singular_by_name import ModbusPointSingularByName
+from src.source_drivers.modbus.resources.point.point_singular import ModbusPointSingularByUUID, \
+    ModbusPointSingularByName
 from src.source_drivers.modbus.resources.point.point_stores import ModbusPointPluralPointStore, ModbusPointStore, \
     ModbusDevicePointPluralPointStore
 from src.system.resources.memory import GetSystemMem
@@ -55,13 +55,13 @@ api_point.add_resource(PointResourceList, '')
 bp_generic = Blueprint('generic', __name__, url_prefix='/api/generic')
 api_generic = Api(bp_generic)
 api_generic.add_resource(GenericNetworkPlural, '/networks')
-api_generic.add_resource(GenericNetworkSingular, '/networks/uuid/<string:uuid>')
+api_generic.add_resource(GenericNetworkSingularByUUID, '/networks/uuid/<string:uuid>')
 api_generic.add_resource(GenericNetworkSingularByName, '/networks/name/<string:name>')
 api_generic.add_resource(GenericDevicePlural, '/devices')
-api_generic.add_resource(GenericDeviceSingular, '/devices/uuid/<string:uuid>')
+api_generic.add_resource(GenericDeviceSingularByUUID, '/devices/uuid/<string:uuid>')
 api_generic.add_resource(GenericDeviceSingularByName, '/devices/name/<string:network_name>/<string:device_name>')
 api_generic.add_resource(GenericPointPlural, '/points')
-api_generic.add_resource(GenericPointSingular, '/points/uuid/<string:uuid>')
+api_generic.add_resource(GenericPointSingularByUUID, '/points/uuid/<string:uuid>')
 api_generic.add_resource(GenericPointSingularByName,
                          '/points/name/<string:network_name>/<string:device_name>/<string:point_name>')
 api_generic.add_resource(GenericUUIDPointValueWriter, '/points_value/uuid/<string:uuid>')
@@ -71,13 +71,13 @@ api_generic.add_resource(GenericNamePointValueWriter,
 bp_modbus = Blueprint('modbus', __name__, url_prefix='/api/modbus')
 api_modbus = Api(bp_modbus)
 api_modbus.add_resource(ModbusNetworkPlural, '/networks')
-api_modbus.add_resource(ModbusNetworkSingular, '/networks/uuid/<string:uuid>')
+api_modbus.add_resource(ModbusNetworkSingularByUUID, '/networks/uuid/<string:uuid>')
 api_modbus.add_resource(ModbusNetworkSingularByName, '/networks/name/<string:name>')
 api_modbus.add_resource(ModbusDevicePlural, '/devices')
-api_modbus.add_resource(ModbusDeviceSingular, '/devices/uuid/<string:uuid>')
+api_modbus.add_resource(ModbusDeviceSingularByUUID, '/devices/uuid/<string:uuid>')
 api_modbus.add_resource(ModbusDeviceSingularByName, '/devices/name/<string:network_name>/<string:device_name>')
 api_modbus.add_resource(ModbusPointPlural, '/points')
-api_modbus.add_resource(ModbusPointSingular, '/points/uuid/<string:uuid>')
+api_modbus.add_resource(ModbusPointSingularByUUID, '/points/uuid/<string:uuid>')
 api_modbus.add_resource(ModbusPointSingularByName,
                         '/points/name/<string:network_name>/<string:device_name>/<string:point_name>')
 api_modbus.add_resource(ModbusPointPoll, '/poll/point/<string:uuid>')
